Detangle amalg., prepare for local imports
